backend/requirements.txt
fastapi==0.135.3
steel-sdk==0.17.0
uvicorn[standard]==0.42.0
playwright==1.58.0
anthropic==0.88.0
